Cleanup Imports and Add X Dummy
All Imports with the form "std::sync::mpsc::Sender" have been cleaned
up to "std::sync::mpsc" and now instead of "Sender", "mpsc::Sender"
would be used. This makes it more clear where methods and structs are
comming from and helps preventing issues with similarily named modules.

The first X Dummy Version is just a proof of concept that an X bar would
work. Everything still has to be done. The only thing working right now
is detecting if Wayland is running and if that is not the case it
launches the Dummy Bar which is currently just an empty Buffer.